LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-waveshapernode-interface/silent-inputs-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "test-0" 
PASS Executing "test-1" 
PASS Executing "test-2" 
PASS Audit report 
PASS > [test-0] curve output is non-zero for silent inputs 
FAIL X WaveShaper with silent inputs and curve {"0":0.5,"1":0.5,"2":0.5}: Expected 0.5 for all values but found 16000 unexpected values: 
	Index	Actual
	[0]	0
	[1]	0
	[2]	0
	[3]	0
	...and 15996 more errors. assert_true: expected true got false
FAIL < [test-0] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [test-1] 2x curve output is non-zero for silent inputs 
FAIL X WaveShaper with 2x oversample, silent inputs, and curve {"0":0.5,"1":0.5,"2":0.5}: Expected 0.5 for all values but found 16000 unexpected values: 
	Index	Actual
	[0]	0
	[1]	0
	[2]	0
	[3]	0
	...and 15996 more errors. assert_true: expected true got false
FAIL < [test-1] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [test-2] curve output is non-zero for no inputs 
FAIL X WaveShaper with no inputs and curve {"0":0.5,"1":0.5,"2":0.5}: Expected 0.5 for all values but found 16000 unexpected values: 
	Index	Actual
	[0]	0
	[1]	0
	[2]	0
	[3]	0
	...and 15996 more errors. assert_true: expected true got false
FAIL < [test-2] 1 out of 1 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 3 out of 3 tasks were failed. assert_true: expected true got false
